+++ app/src/main/java/com/perno97/financialmanagement/database/Profile.kt
@@ -9,7 +9,7 @@ import java.time.LocalDate
 
 @Entity
 data class Profile(
-    @PrimaryKey val profileId: Int,
+    @PrimaryKey val profileId: Long,
     val assets: Float,
     @ColumnInfo(name = "last_access") val lastAccess: LocalDate
 )
