+++ app/src/main/java/com/perno97/financialmanagement/utils/MovementDetailsData.kt
@@ -4,17 +4,17 @@ import java.time.DayOfWeek
 import java.time.LocalDate
 
 data class MovementDetailsData(
-    val movementId: Int?,
+    val movementId: Long?,
     val date: LocalDate,
     val amount: Float,
     val category: String,
     val color: String,
     val title: String,
     val notes: String,
-    val periodicMovementId: Int?,
+    val periodicMovementId: Long?,
     val weekDays: List<DayOfWeek>?,
     val days: Int,
     val months: Int,
     val notify: Boolean,
-    val incomingMovementId: Int?
+    val incomingMovementId: Long?
 )
